+++ netsci/tests/metrics/profile_motifs_gpu.py
@@ -1,52 +1,45 @@
-import logging; logging.basicConfig()
+import logging
 
-import tensorflow as tf
-logging.getLogger('netsci.metrics.motifs_gpu').setLevel(logging.DEBUG)
-
+from netsci.system_info import get_system_info
 from netsci.models.random import erdos_renyi
 import netsci.metrics.motifs as nsm
 from netsci.tests.profiler import Timer
 
-from tensorflow.python.client import device_lib
+logging.basicConfig()
+# logging.getLogger("netsci.metrics.motifs_gpu").setLevel(logging.DEBUG)
 
 
-print('devices:', {d.name: d.physical_device_desc 
-                   for d in device_lib.list_local_devices()})
+print(get_system_info()["gpus_compute"])
 
 
 A = erdos_renyi(n=2000, p=0.1, random_state=71070)
-print(f'A: {A.shape}, p = {nsm.sparsity(A):.4f}')
+print(f"A: {A.shape}, p = {nsm.sparsity(A):.4f}")
 
-configurations = [('matmul', 'GPU'), ('matmul', 'CPU'), ('louzoun', 'CPU')]
+for algorithm in ["louzoun"]:
+    with Timer(logger=print, desc=f"{algorithm}"):
+        print(nsm.motifs(A, algorithm=algorithm))
 
-for algorithm, device in configurations:
-  with tf.device(f'/{device}:0'):
-    with Timer(logger=print, desc=f'{algorithm}:{device}'):
-      print(nsm.motifs(A, algorithm=algorithm))
+for tensorian, _ in nsm.motifs_matmul.tensorians.items():
+    with Timer(logger=print, desc=f"{tensorian}"):
+        print(nsm.motifs_matmul.motifs_matmul(A, tensorian))
 
 
-"""Colab
-devices = {
-  '/device:CPU:0': '', 
-  '/device:GPU:0': 'device: 0, name: Tesla K80, pci bus id: 0000:00:04.0, compute capability: 3.7'
-}
-
+"""Tesla T4 @ Colab
 A: (2000, 2000), p = 0.1001
 
-matmul:GPU took 0.680 sec.
-matmul:CPU took 18.909 sec.
-louzoun:CPU took 149.726 sec.
+louzoun took 127.716 sec.
+numpy-cpu took 9.627 sec.
+tensorflow-cpu took 0.626 sec.
+tensorflow-gpu took 0.129 sec.
+torch-cpu took 4.698 sec.
+torch-gpu took 0.390 sec.
 """
 
-"""MBP M1 Max
-devices: {
-  '/device:CPU:0': '', 
-  '/device:GPU:0': 'device: 0, name: METAL, pci bus id: <undefined>'
-}
-
+"""MBP M3 Max
 A: (2000, 2000), p = 0.1001
 
-matmul:GPU took 1.943 sec.
-matmul:CPU took 1.868 sec.
-louzoun:CPU took 64.622 sec.
+louzoun took 47.796 sec.
+numpy-cpu took 0.690 sec.
+torch-cpu took 0.209 sec.
+torch-gpu took 0.168 sec.
 """